Fix compile warnings on Windows
C:\Buildbot\gammaray-qt57-msvc2015-win64-windows\source\probe\abstractfunctionoverwriter.cpp(58):
	 warning C4311: 'reinterpret_cast': pointer truncation from 'void *const ' to 'unsigned long'
C:\Buildbot\gammaray-qt57-msvc2015-win64-windows\source\probe\abstractfunctionoverwriter.cpp(58):
	 warning C4302: 'reinterpret_cast': truncation from 'void *const ' to 'unsigned long'
C:\Buildbot\gammaray-qt57-msvc2015-win64-windows\source\probe\abstractfunctionoverwriter.cpp(59):
	 warning C4311: 'reinterpret_cast': pointer truncation from 'quint8 *' to 'unsigned long'
C:\Buildbot\gammaray-qt57-msvc2015-win64-windows\source\probe\abstractfunctionoverwriter.cpp(59):
	 warning C4302: 'reinterpret_cast': truncation from 'quint8 *' to 'unsigned long'

It is questionable, whether the left-hand side is valid though, as it
stores the potentially 64bit large ptrdiff in a quint32... So
hopefully the short jump distance is really less then 32bit.

+++ probe/abstractfunctionoverwriter.cpp
@@ -55,8 +55,8 @@ bool AbstractFunctionOverwriter::writeShortJump(void *target, void * const func)
 
     *cur = 0xE9;
     cur++;
-    *reinterpret_cast<quint32 *>(cur) = reinterpret_cast<unsigned long>(func)
-                                        - reinterpret_cast<unsigned long>(cur + 4);
+    *reinterpret_cast<quint32 *>(cur) = reinterpret_cast<uintptr_t>(func)
+                                        - reinterpret_cast<uintptr_t>(cur + 4);
 
     ret = reprotectMemory(page_align(target), roundToNextPage(5));
